refactor: Split the Twitch controller, and stop replaying signed deliveries
controller/twitch.py held webhook transport, seven route stubs, eight event
tasks and two OAuth flows. A dependency scan showed three of those four groups
call nothing else in the file; only the route stubs reach across, to the
dispatcher and to a handler, which is what a route stub is for. The seams were
already there, undeclared.

services/twitch/events.py takes the event tasks. They reference no HTTP name at
all - no Request, no Response, no HTTPException - so they were never controller
code; they orchestrate the alert, the session, chat and Helix and merely happen
to be triggered by a webhook. Nothing under services/ imports controller/, so
the direction stays one way. The handlers lose their leading underscore, being
the module's interface now rather than a file-local detail.

controller/twitch_oauth.py takes the two authorization-code flows on their own
router. It shares nothing with the webhook controller: not the signature check,
not the models, not the path prefix.

602 -> 323, 206 and 202. One move alone would not have cleared the threshold:
OAuth only reaches ~433, the tasks only ~437.

Replay protection, which the gate blocked on and which the user chose. A
delivery whose timestamp is more than ten minutes from now in either direction
is refused - a clock ahead is as wrong as one behind. A message id that already
reached a handler answers 202 without dispatching again, because Twitch says a
notification may arrive twice and 202 is what stops the retries. Claiming and
checking are one step, since there is an await before the dispatch and two
copies of one delivery could otherwise both pass a check that only looked; a
delivery that then fails releases its claim, so what Twitch retries because this
end failed still gets through. The cost is stated in AGENTS.md rather than
hidden: a host clock more than ten minutes out refuses every event, loudly.

Verified by snapshotting the registered route table and driving a signed
stream.online payload at all seven webhook routes, before the split and after.
Paths, methods and route names identical; the 202 still lands on
/webhook/twitch and the other six still answer 400, which is what would move had
a model been rewired. The only differences are the three OAuth endpoints' module
names. Ten replay cases pass, including the one that matters: a delivery that
failed validation is retried and still dispatches.

Two harnesses had to be fixed to keep telling the truth - they signed with a
hardcoded 2025 timestamp, now stale by design, and reused one message id across
all seven routes, which deduplication would have read as duplicates.

Also on the way past: the verification challenge is echoed as text/plain; the
lifespan handler gained its return annotation; and an OAuth denial reason, which
arrives in the callback query string and is nobody's word for anything, is
capped, logged as a repr so a newline cannot forge a line, escaped, and sent to
the admin channel instead of reflected to the caller.
