+++ controller/__init__.py
@@ -1,3 +1,4 @@
 from .twitch import twitch_router
+from .twitch_oauth import twitch_oauth_router
 
-__all__ = ["twitch_router"]
+__all__ = ["twitch_oauth_router", "twitch_router"]
