+++ controller/twitch.py
@@ -1,11 +1,11 @@
-import asyncio
 import logging
 import time
+from collections import OrderedDict
 from collections.abc import Callable, Coroutine
 from typing import Any
 
+import pendulum
 from fastapi import APIRouter, HTTPException, Request, Response
-from fastapi.responses import RedirectResponse
 from pydantic import BaseModel, ValidationError
 
 from background import fire_and_forget
@@ -16,7 +16,6 @@ from constants import (
     TWITCH_MESSAGE_SIGNATURE,
     TWITCH_MESSAGE_TIMESTAMP,
     TWITCH_MESSAGE_TYPE,
-    TokenType,
 )
 from errors import notify, report
 from models import (
@@ -25,47 +24,67 @@ from models import (
     ChannelFollowEventSub,
     ChannelModerateEventSub,
     ChannelRaidEventSub,
-    RefreshResponse,
-    Stream,
     StreamOfflineEventSub,
     StreamOnlineEventSub,
-    TokenValidationResponse,
 )
 from services import (
     get_hmac,
     get_hmac_message,
-    get_stream,
-    get_user,
+    parse_rfc3339,
     verify_message,
 )
-from services.config import config
-from services.helper.http_client import http_client_manager
-from services.twitch import live_alert, stream_session
-from services.twitch.chat import say, say_template
-from services.twitch.commands import dispatch
-from services.twitch.helix import HelixError
-from services.twitch.oauth import (
-    authorization_url,
-    callback_uri,
-    configured_scopes,
-    consume_authorization,
-    expected_user_id,
-)
-from services.twitch.token_manager import token_manager
+from services.twitch import events
 
 logger = logging.getLogger(__name__)
 
 twitch_router = APIRouter()
 
-# Twitch announces the stream before Helix lists it. Bounded by the clock rather
-# than by a count of attempts: a lookup that is timing out takes most of a minute
-# on its own, so thirty of those is a quarter of an hour, not thirty seconds.
-_STREAM_WAIT_SECONDS = 60
 
 # An EventSub payload is a few kilobytes. The whole body has to be read to
 # compute the signature over it, so it is counted as it arrives and abandoned
 # past this.
 _MAX_BODY_BYTES = 256 * 1024
+
+
+# Twitch's own guidance for replay: a delivery whose timestamp is far from now
+# is not one Twitch is sending, it is one somebody kept. Applied in both
+# directions, since a clock ahead is as wrong as a clock behind.
+_MESSAGE_WINDOW_SECONDS = 600
+
+# Ids of deliveries that reached a handler. Only those, so a delivery Twitch
+# retries because this end failed still gets through, while a retry of one that
+# worked does not run it twice. Process-local is enough and not a compromise:
+# the Discord gateway connection lives in this process, so a second replica
+# would double every bot action.
+_HANDLED_LIMIT = 1024
+_handled: OrderedDict[str, float] = OrderedDict()
+
+
+def _is_stale(sent: pendulum.DateTime) -> bool:
+    return abs((pendulum.now("UTC") - sent).total_seconds()) > _MESSAGE_WINDOW_SECONDS
+
+
+def _claim(message_id: str) -> bool:
+    """Take this delivery, or say that something already has it.
+
+    Claiming and checking are one step: there is an await between here and the
+    dispatch, so two copies of one delivery could otherwise both get past a
+    check that only looked.
+    """
+    cutoff = time.monotonic() - _MESSAGE_WINDOW_SECONDS
+    while _handled and next(iter(_handled.values())) < cutoff:
+        _handled.popitem(last=False)
+    if message_id in _handled:
+        return False
+    _handled[message_id] = time.monotonic()
+    while len(_handled) > _HANDLED_LIMIT:
+        _handled.popitem(last=False)
+    return True
+
+
+def _release(message_id: str) -> None:
+    """Give a claim back, so a delivery this end failed can still be retried."""
+    _handled.pop(message_id, None)
 
 
 async def _bounded_body(request: Request, endpoint: str) -> bytes:
@@ -123,6 +142,25 @@ async def validate_call(request: Request, endpoint: str) -> Response | None:
         await notify(f"403: Forbidden request on {endpoint}. Signature does not match.")
         raise HTTPException(status_code=403)
 
+    # Only after the signature matched: an unauthenticated caller must not be
+    # able to raise a notice by sending a stale timestamp.
+    try:
+        sent = parse_rfc3339(headers.get(TWITCH_MESSAGE_TIMESTAMP, ""))
+    except ValueError:
+        logger.warning("403: Unreadable message timestamp on %s", endpoint)
+        await notify(f"403: Forbidden request on {endpoint}. Unreadable timestamp.")
+        raise HTTPException(status_code=403) from None
+
+    if _is_stale(sent):
+        logger.warning("403: Stale delivery on %s, sent %s", endpoint, sent)
+        await notify(
+            f"403: Forbidden request on {endpoint}. Timestamp outside the"
+            f" {_MESSAGE_WINDOW_SECONDS}s window, which is also what a wrong"
+            f" clock on this host looks like.",
+            key=f"stale-delivery:{endpoint}",
+        )
+        raise HTTPException(status_code=403)
+
     message_type = headers.get(TWITCH_MESSAGE_TYPE, "").lower()
     try:
         parsed = await request.json()
@@ -142,8 +180,12 @@ async def validate_call(request: Request, endpoint: str) -> Response | None:
 
     if message_type == "webhook_callback_verification":
         challenge = body.get("challenge")
+        # text/plain explicitly: the value is echoed straight back, and nothing
+        # that reads the reply should be free to interpret it as markup.
         return Response(
-            challenge if isinstance(challenge, str) else "", status_code=200
+            challenge if isinstance(challenge, str) else "",
+            status_code=200,
+            media_type="text/plain",
         )
 
     if message_type == "revocation":
@@ -175,9 +217,22 @@ async def process_webhook[E: BaseModel](
         if validation:
             return validation
 
-        body: dict[str, Any] = await request.json()
-        event_sub = event_model.model_validate(body)
-        fire_and_forget(task_func(event_sub), name=endpoint)
+        # Twitch says a notification may arrive twice. 202 rather than an error:
+        # this end does have it, and saying so is what stops the retries.
+        message_id = request.headers.get(TWITCH_MESSAGE_ID, "")
+        if not _claim(message_id):
+            logger.info(
+                "Duplicate %s on %s, not dispatched again", message_id, endpoint
+            )
+            return Response(status_code=202)
+
+        try:
+            body: dict[str, Any] = await request.json()
+            event_sub = event_model.model_validate(body)
+            fire_and_forget(task_func(event_sub), name=endpoint)
+        except BaseException:
+            _release(message_id)
+            raise
         return Response(status_code=202)
     except HTTPException:
         raise
@@ -198,314 +253,13 @@ async def process_webhook[E: BaseModel](
         raise HTTPException(status_code=500) from e
 
 
-async def _wait_for_stream_info(
-    broadcaster_id: int,
-) -> tuple[Stream | None, HelixError | None]:
-    """Poll until Twitch admits the stream is up, tolerating a failed lookup.
-
-    Bounded, because Twitch delivers stream.online once and never again: waiting
-    forever and giving up both lose the alert, but only one of them says so. The
-    last lookup error comes back too, so giving up can tell "Twitch says they
-    are offline" from "Twitch could not be reached".
-    """
-    deadline = time.monotonic() + _STREAM_WAIT_SECONDS
-    last_error: HelixError | None = None
-
-    while True:
-        try:
-            stream_info = await get_stream(broadcaster_id)
-        except HelixError as e:
-            last_error = e
-            logger.warning(f"Stream lookup failed for {broadcaster_id}: {e}")
-            stream_info = None
-
-        if stream_info:
-            return stream_info, None
-        # Checked after the attempt, so a slow first lookup still gets its turn.
-        if time.monotonic() >= deadline:
-            return None, last_error
-        await asyncio.sleep(1)
-
-
-async def _stream_online_task(event_sub: StreamOnlineEventSub) -> None:
-    broadcaster_id = int(event_sub.event.broadcaster_user_id)
-    try:
-        stream_info, lookup_error = await _wait_for_stream_info(broadcaster_id)
-        if stream_info is None:
-            reason = (
-                f"the last lookup failed: {lookup_error}"
-                if lookup_error is not None
-                else "Twitch went on reporting them offline"
-            )
-            await notify(
-                f"Gave up after {_STREAM_WAIT_SECONDS}s waiting for Twitch to confirm"
-                f" broadcaster {broadcaster_id} is live, so no alert was posted and no"
-                f" stream session was started: {reason}",
-                key=f"stream-online-gave-up:{broadcaster_id}",
-            )
-            return
-
-        try:
-            user_info = await get_user(broadcaster_id)
-        except HelixError as e:
-            # Decoration. stream.online does not come again, so the alert must
-            # not depend on it.
-            await notify(
-                f"Announcing broadcaster {broadcaster_id} without their profile: {e}",
-                key=f"stream-online-profile:{broadcaster_id}",
-            )
-            user_info = None
-
-        is_main = stream_info.user_login == config.setting("broadcaster_username")
-        channel = (
-            config.channel("stream_alerts") if is_main else config.channel("promo")
-        )
-
-        # began does not raise: it guards each of its own chat lines, because
-        # the alert below must survive a Twitch that will not take a greeting.
-        await stream_session.began(broadcaster_id, stream_info)
-
-        await live_alert.announce(broadcaster_id, stream_info, user_info, channel)
-
-    except Exception as e:  # noqa: BLE001
-        await report(e, f"Error in _stream_online_task for {broadcaster_id}")
-
-
-async def _stream_offline_task(event_sub: StreamOfflineEventSub) -> None:
-    broadcaster_id = int(event_sub.event.broadcaster_user_id)
-    try:
-        # The payload names no stream, so this handler cannot tell which one
-        # ended. It wakes the updater, which can. See docs/adr/0001.
-        await live_alert.wake(broadcaster_id)
-
-    except Exception as e:  # noqa: BLE001
-        await report(e, f"Error in _stream_offline_task for {broadcaster_id}")
-
-
-async def _channel_chat_message_task(event_sub: ChannelChatMessageEventSub) -> None:
-    try:
-        if not event_sub.event.message.text.startswith("!"):
-            return
-        text_without_prefix = event_sub.event.message.text[1:]
-        command_parts = text_without_prefix.split(" ", 1)
-        command = command_parts[0].lower()
-        args = command_parts[1] if len(command_parts) > 1 else ""
-
-        if (
-            event_sub.event.source_broadcaster_user_id is not None
-            and event_sub.event.source_broadcaster_user_id
-            != event_sub.event.broadcaster_user_id
-        ):
-            return
-
-        await dispatch(event_sub, command, args)
-    except Exception as e:  # noqa: BLE001
-        await report(e, "Error processing Twitch chat webhook task")
-
-
-async def _channel_follow_task(event_sub: ChannelFollowEventSub) -> None:
-    try:
-        await say_template(
-            event_sub.event.broadcaster_user_id,
-            "twitch_follow_thanks",
-            user=event_sub.event.user_name,
-        )
-    except Exception as e:  # noqa: BLE001
-        await report(e, "Error processing Twitch follow webhook task")
-
-
-async def _channel_ad_break_begin_task(event_sub: ChannelAdBreakBeginEventSub) -> None:
-    broadcaster_id = event_sub.event.broadcaster_user_id
-    try:
-        ad_duration = event_sub.event.duration_seconds
-
-        # Each step stands alone: a dropped "ads starting" used to take the
-        # "ads over" message and the next break's warning down with it.
-        await say_template(
-            broadcaster_id, "twitch_ad_break_start", minutes=ad_duration // 60
-        )
-        await asyncio.sleep(ad_duration)
-        await say_template(broadcaster_id, "twitch_ad_break_end")
-
-        stream_session.schedule_ad_break_warning(broadcaster_id)
-    except Exception as e:  # noqa: BLE001
-        await report(e, "Error processing Twitch ad break webhook task")
-
-
-async def _validate_oauth_identity(
-    auth_response: RefreshResponse,
-    token_type: TokenType,
-    endpoint: str,
-) -> TokenValidationResponse:
-    response = await http_client_manager.request(
-        "GET",
-        "https://id.twitch.tv/oauth2/validate",
-        headers={"Authorization": f"OAuth {auth_response.access_token}"},
-    )
-    if response.status_code < 200 or response.status_code >= 300:
-        logger.error(
-            "OAuth token validation failed with status=%s", response.status_code
-        )
-        await notify(
-            f"Failed to validate the Twitch token from {endpoint}: "
-            f"{response.status_code} {response.text}"
-        )
-        raise HTTPException(status_code=500, detail="Twitch token validation failed")
-
-    validation = TokenValidationResponse.model_validate(response.json())
-    expected_id = expected_user_id(token_type)
-    missing_scopes = sorted(set(configured_scopes()) - set(validation.scopes))
-
-    problems = []
-    if validation.client_id != settings.twitch_client_id:
-        problems.append("the token belongs to a different Twitch application")
-    if validation.user_id != expected_id:
-        problems.append(
-            f"expected Twitch user ID {expected_id}, but {validation.login} "
-            f"authorized as ID {validation.user_id}"
-        )
-    if missing_scopes:
-        problems.append(f"missing scopes: {', '.join(missing_scopes)}")
-
-    if problems:
-        message = f"Rejected {token_type.value} authorization: {'; '.join(problems)}"
-        logger.warning(message)
-        await notify(message)
-        raise HTTPException(status_code=400, detail=message)
-    return validation
-
-
-async def _oauth_callback_common(
-    code: str | None,
-    state: str,
-    endpoint: str,
-    token_type: TokenType,
-    error: str | None,
-    error_description: str | None,
-) -> tuple[RefreshResponse, TokenValidationResponse]:
-    if not consume_authorization(token_type, state):
-        logger.warning("400: Bad request on %s. Invalid OAuth state.", endpoint)
-        await notify(f"400: Bad request on {endpoint}. Invalid or expired OAuth state.")
-        raise HTTPException(status_code=400, detail="Invalid or expired OAuth state")
-
-    if error:
-        detail = error_description or error
-        logger.warning("Twitch authorization denied on %s: %s", endpoint, detail)
-        raise HTTPException(
-            status_code=400, detail=f"Twitch authorization denied: {detail}"
-        )
-    if not code:
-        raise HTTPException(status_code=400, detail="Missing Twitch authorization code")
-
-    params = {
-        "client_id": settings.twitch_client_id,
-        "client_secret": settings.twitch_client_secret,
-        "code": code,
-        "grant_type": "authorization_code",
-        "redirect_uri": callback_uri(token_type),
-    }
-    response = await http_client_manager.request(
-        "POST", "https://id.twitch.tv/oauth2/token", data=params
-    )
-
-    if response.status_code < 200 or response.status_code >= 300:
-        logger.error(
-            f"Token exchange failed with status={response.status_code}, response={response.text}"
-        )
-        await notify(
-            f"Failed to exchange token: {response.status_code} {response.text}"
-        )
-        raise HTTPException(status_code=500)
-
-    auth_response = RefreshResponse.model_validate(response.json())
-    if auth_response.token_type != "bearer":
-        logger.error(
-            f"Token exchange failed: unexpected token type {auth_response.token_type}"
-        )
-        await notify(
-            f"Failed to exchange token: unexpected token type {auth_response.token_type}"
-        )
-        raise HTTPException(status_code=500)
-    validation = await _validate_oauth_identity(auth_response, token_type, endpoint)
-    return auth_response, validation
-
-
-@twitch_router.get("/twitch/oauth/start/{identity}")
-async def twitch_oauth_start(identity: str, state: str) -> Response:
-    try:
-        token_type = TokenType(identity)
-        return RedirectResponse(authorization_url(token_type, state), status_code=302)
-    except ValueError as e:
-        raise HTTPException(status_code=400, detail=str(e)) from None
-
-
-@twitch_router.get("/twitch/oauth/callback")
-async def twitch_oauth_callback(
-    state: str,
-    code: str | None = None,
-    error: str | None = None,
-    error_description: str | None = None,
-) -> Response:
-    try:
-        auth_response, validation = await _oauth_callback_common(
-            code,
-            state,
-            "/twitch/oauth/callback",
-            TokenType.User,
-            error,
-            error_description,
-        )
-        await token_manager.set_user_access_token(auth_response)
-        return Response(
-            f"Authorization successful for bot account {validation.login} "
-            f"({validation.user_id}). You can close this tab.",
-            status_code=200,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        await report(e, "500: Internal server error on /twitch/oauth/callback")
-        raise HTTPException(status_code=500) from e
-
-
-@twitch_router.get("/twitch/oauth/callback/broadcaster")
-async def twitch_oauth_callback_broadcaster(
-    state: str,
-    code: str | None = None,
-    error: str | None = None,
-    error_description: str | None = None,
-) -> Response:
-    try:
-        auth_response, validation = await _oauth_callback_common(
-            code,
-            state,
-            "/twitch/oauth/callback/broadcaster",
-            TokenType.Broadcaster,
-            error,
-            error_description,
-        )
-        await token_manager.set_broadcaster_access_token(auth_response)
-        return Response(
-            f"Authorization successful for broadcaster account {validation.login} "
-            f"({validation.user_id}). You can close this tab.",
-            status_code=200,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        await report(
-            e, "500: Internal server error on /twitch/oauth/callback/broadcaster"
-        )
-        raise HTTPException(status_code=500) from e
-
-
 @twitch_router.post("/webhook/twitch")
 async def stream_online_webhook(request: Request) -> Response:
     return await process_webhook(
         request,
         "/webhook/twitch",
         StreamOnlineEventSub,
-        _stream_online_task,
+        events.stream_online,
     )
 
 
@@ -515,7 +269,7 @@ async def stream_offline_webhook(request: Request) -> Response:
         request,
         "/webhook/twitch/offline",
         StreamOfflineEventSub,
-        _stream_offline_task,
+        events.stream_offline,
     )
 
 
@@ -525,7 +279,7 @@ async def channel_chat_message_webhook(request: Request) -> Response:
         request,
         "/webhook/twitch/chat",
         ChannelChatMessageEventSub,
-        _channel_chat_message_task,
+        events.channel_chat_message,
     )
 
 
@@ -535,7 +289,7 @@ async def channel_follow_webhook(request: Request) -> Response:
         request,
         "/webhook/twitch/follow",
         ChannelFollowEventSub,
-        _channel_follow_task,
+        events.channel_follow,
     )
 
 
@@ -545,30 +299,8 @@ async def channel_ad_break_begin_webhook(request: Request) -> Response:
         request,
         "/webhook/twitch/adbreak",
         ChannelAdBreakBeginEventSub,
-        _channel_ad_break_begin_task,
+        events.channel_ad_break_begin,
     )
-
-
-async def _channel_raid_task(event_sub: ChannelRaidEventSub) -> None:
-    try:
-        if stream_session.is_main_broadcaster(event_sub.event.from_broadcaster_user_id):
-            twitch_url = (
-                f"https://www.twitch.tv/{event_sub.event.to_broadcaster_user_login}"
-            )
-            await say_template(
-                event_sub.event.from_broadcaster_user_id,
-                "twitch_raid_out",
-                name=event_sub.event.to_broadcaster_user_name,
-                url=twitch_url,
-            )
-        elif stream_session.is_main_broadcaster(event_sub.event.to_broadcaster_user_id):
-            await say(
-                event_sub.event.to_broadcaster_user_id,
-                f"!so {event_sub.event.from_broadcaster_user_login}",
-                "the incoming raid shoutout",
-            )
-    except Exception as e:  # noqa: BLE001
-        await report(e, "Error processing Twitch raid webhook task")
 
 
 @twitch_router.post("/webhook/twitch/raid")
@@ -577,19 +309,8 @@ async def channel_raid_webhook(request: Request) -> Response:
         request,
         "/webhook/twitch/raid",
         ChannelRaidEventSub,
-        _channel_raid_task,
+        events.channel_raid,
     )
-
-
-async def _channel_moderate_task(event_sub: ChannelModerateEventSub) -> None:
-    try:
-        if event_sub.event.action != "raid" or not stream_session.is_main_broadcaster(
-            event_sub.event.broadcaster_user_id
-        ):
-            return
-        await say_template(event_sub.event.broadcaster_user_id, "twitch_raid_farewell")
-    except Exception as e:  # noqa: BLE001
-        await report(e, "Error processing Twitch moderate webhook task")
 
 
 @twitch_router.post("/webhook/twitch/moderate")
@@ -598,5 +319,5 @@ async def channel_moderate_webhook(request: Request) -> Response:
         request,
         "/webhook/twitch/moderate",
         ChannelModerateEventSub,
-        _channel_moderate_task,
+        events.channel_moderate,
     )
